requirements.txt
spotipy
pandas
python-dotenv
playwright